+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Jongwhan Lee (@leejw51crypto) ([#878]).
 
 - [ibc-relayer-cli]
   - Fix for `ft-transfer` mismatching arguments ([#869])
+  - Fix channel destination chain mismatch on unreceived-packets or unreceived-acks ([#873])
 
 ### BREAKING CHANGES
 
@@ -57,7 +58,7 @@ Jongwhan Lee (@leejw51crypto) ([#878]).
 [#863]: https://github.com/informalsystems/ibc-rs/issues/863
 [#869]: https://github.com/informalsystems/ibc-rs/issues/869
 [#878]: https://github.com/informalsystems/ibc-rs/issues/878
-
+[#873]: https://github.com/informalsystems/ibc-rs/issues/873
 
 ## v0.2.0
 *April 14th, 2021*